Serialize `Profile.retrieve` instead of `Profile.fetch`
The former contains also the information about the stackframes and can be used
for analysis in a different session with something like

```
Profile.print(deserialize("/path/to/profile_file.dat")...)
```

or similarly for other profile analysis tools.

NOTE: since this is using `Serialization.serialize` you must deserialize the
data with the same version of Julia as the one used for serialization.